+++ game-h5/game-h5-controller/src/main/java/com/fulu/game/h5/controller/mp/MpOrderController.java
@@ -64,7 +64,7 @@ public class MpOrderController extends BaseController {
     @PostMapping("/submit")
     public Result submit(HttpServletRequest request,
                          @RequestParam String sessionkey,
-                         @RequestParam Integer virtualMoney,
+                         @RequestParam Long virtualMoney,
                          @RequestParam Integer payment) {
 
         User user = userService.getCurrentUser();
